requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
Django==4.1
djangorestframework==3.13.1
idna==3.3
python-dotenv==0.20.0
pytz==2022.2
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
uuid==1.30
